requirements.txt
discord.py==2.4.0
python-dotenv>=1.0.1
cryptography>=42.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0